Extract preparedCheckSync service to address reusability for both editing or deleting access arrangements
